IMPALA-12372: Only use -Wno-deprecated-declaration for OpenSSL3
Redhat 9 and Ubuntu 22.04 both use OpenSSL3, which deprecated
several APIs that we use. To support those platforms, we added
the -Wno-deprecated-declaration to the build. Historically, the
Impala build has also specified -Wno-deprecated due to
use of deprecated headers in gutils. These flags limit our
ability to notice use of deprecated code in other parts of the
code.

The code in gutils no longer requires -Wno-deprecated, so
this removes it completely. Additionally, this limits the
-Wno-deprecated-declaration flag to machines using
OpenSSL 3.

Reenabling deprecation warnings also reenables Clang Tidy's
clang-diagnostic-deprecated enforcement. This is currently
broken, so this turns off clang-diagnostic-deprecated
until it can be addressed properly.

Testing:
 - Ran build-all-options on Ubuntu 22 and Ubuntu 16
 - Ran a Rocky 9.2 build

Change-Id: I1b36450d084f342eeab5dac2272580ab6b0c988b
Reviewed-on: http://gerrit.cloudera.org:8080/20369
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>