IMPALA-9118: Show catalog operation details in catalogd webUI
This patch extends the /operations page in catalogd WebUI to show the
in-flight and finished catalog operations. The following fields are
shown for each operation:
 - Thread ID
 - Query ID
 - Client IP
 - Coordinator
 - Type
 - Target
 - User
 - Start Time
 - End Time (only shown for finished operations)
 - Duration
 - Status
 - Details

Finished operation records are currently kept in memory and limited by
the catalog_operation_log_size flag (defaults to 100).

To collect the above fields, this patch extends
TCatalogServiceRequestHeader to contain the coordinator hostname. Also
fix some catalog RPCs that didn't fill these fields.

Tests:
 - Add e2e test in custom_cluster/test_web_pages.py
 - Manually verify the web pages when running a GVO job

Change-Id: I3cf3f0da2be2be79e546762a8083d4de338ff6aa
Reviewed-on: http://gerrit.cloudera.org:8080/20428
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>