IMPALA-12432: Make LdapKerberosImpalaShellTest* work with Guava 28
In the change for IMPALA-11726, we added some usages of
Guava's ImmutableMap.of() with >5 key/value pairs. This
won't compile in older versions of Guava (like Guava 28)
that only support up to 5 key/value pairs in ImmutableMap.of().

Being compatible with older versions of Guava makes it
easier for Impala to match the Guava version of Hadoop,
Hive, etc when compiling against older versions.

This switches to using ImmutableMap.builder() to build
the larger maps.

Testing:
 - Built with Guava 28 and ran the tests

Change-Id: Ic6d9c69fff749bcdb0887a6676574db9526f68e8
Reviewed-on: http://gerrit.cloudera.org:8080/20471
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>