IMPALA-12411: Fix data race in expr-test teardown
TSAN ThreadSanitizer detected a data race during expr-test teardown when
a query finishes after JvmMetricCache's destructor is called. expr-test
can finish with a query still running because ImpaladQueryExecutor
closes the previous query when you start the next query and relies on
the destructor to close the last query.

expr-test creates ImpaladQueryExecutor as a global as part of setting up
an in-memory cluster (with Statestore and InProcessImpalaServer). So the
last query isn't guaranteed to be closed until global destruction, which
leads to a race with other global destruction (in this case with
JvmMetricCache while updating QueryState on query completion).

ImpaladQueryExecutor is only used in expr-test.

Change-Id: I8e289b292cb11154f1245e2b987cde0995b6243c
Reviewed-on: http://gerrit.cloudera.org:8080/20474
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>