IMPALA-10949: Improve batching logic of partition events
Currently the batching logic for partitions evaluates self-event check
by looking at the event id of the last event in the batch. This patch
improves the batching logic by evaluating table's lastSyncEventId to
the current event's event id and decide whether to batch the event or
not. This way we can have fewer batch sizes and avoid self events if
any, beforehand.

Testing: Added an end-to-end test to verfiy the batch size for
ALTER_PARTITION batch events.

Change-Id: I4e79510739347cbe669719a9e4cb9cabd5daa7d3
Reviewed-on: http://gerrit.cloudera.org:8080/20485
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>