IMPALA-12399: (addendum) Fixed possible deadloop and also do not set the event skip list in notification request to update latestEventId_
Following points are fixed:
1. Possible deadloop in
MetastoreEventsProcessor#getNextMetastoreEventsInBatches()
when OPEN_TXN is the last event on HMS at the moment the
method is invoked. Need to return when events received
in response is empty.
2. latestEventId_ reflects the latest event in HMS. Should
not set the event skip list in notification request to
update latestEventId_.
3. Update lastSyncedEventId_ when event list returned is
empty due to event skip list in notification event.

Testing:
- Added tests and executed all existing tests.

Change-Id: Idb4b8c3db23d39226f10b33cca4e6a1ab271b925
Reviewed-on: http://gerrit.cloudera.org:8080/20487
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>