IMPALA-12453: Increase yarn high disk threshold
Yarn will mark a node as unavailable for running jobs if it exceeds the
disk threshold. We only have one node, so that means jobs hang until
disk frees up. Exhaustive test runs are approaching using 300GB, which
is a common size for our test environments, and we see some tests
hanging on yarn jobs.

Increases the disk threshold from default of 90% to 99%; in test
environments the remaining 1% is ~3GB. That should still be sufficient
in case yarn needs to throttle workloads to avoid using all the disk,
while leaving most of the disk available.

Change-Id: I06b723e4ecea94ce95fb2bd41bd356fba694e87a
Reviewed-on: http://gerrit.cloudera.org:8080/20501
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>