IMPALA-12398: Fix Ranger role not exists when altering db/table/view owner to a role
When Role '<ROLE_NAME>' is created with Ranger authorization enabled,
if 'ALTER TABLE <TABLE_NAME> SET OWNER ROLE <ROLE_NAME>' statement is
executed to assign role as the owner of the table, it will throw
AnalysisException:Role '<ROLE_NAME>' does not exist.

Before this patch, given the ALTER DATABASE/TABLE/VIEW SET OWNER ROLE
statement, Impala always checked the existence of the given role in
its AuthorizationPolicy. However, when the support for role-related
statements with Ranger was added in IMPALA-10211, we only added the
roles in RangerImpalaPlugin instead of AuthorizationPolicy.
Therefore, the statement above  would fail even though an authorized
user tries to set the owner to an existing role in RangerImpalaPlugin.

This patch will directly use ranger impala plugin to check the
existence of the role, instead of using AuthorizationPolicy object.

Tests:
 - Pass unit tests. test method testAlterView in AuthorizationStmtTest
   is updated accordingly.
 - Pass e2e tests. test method _test_ownership in test_ranger.py is
   updated to cover the new implementation.
 - Pass core tests with ranger enabled.

Change-Id: I2b029bdb90111dbd0eab5189360cc81090225cda
Reviewed-on: http://gerrit.cloudera.org:8080/20508
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>