IMPALA-12461 part1: Avoid taking db/table level locks db/table self-event check
DB / Table level locks can be held for a long time by DDL/DML
operations in the catalogd. Trying to get the lock during self-event
check would mean blocking if there is an ongoing operation for
the given db/table.

This patch tries to solve it for the easy case of db/table level
events by using self-event specific locking that should be only
taken for short times to add/remove events from/to the in-flight
event list. The InFlightEvents object itself is used as lock
and no other of catalogd's lock must be acquired after it to
avoid deadlock.

Postponing solving this for partition level events as that would
be more complex, as both the partition list and the partitions'
in-flight event lists would need to be protected from parallel
operations that add/remove partitions.

Testing:
- ran event processing related tests

Change-Id: Ife455de09ab2e262bde1e4b5bd54c8c54c75f2cd
Reviewed-on: http://gerrit.cloudera.org:8080/20516
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>