IMPALA-12478: Invalid length in StringValue::LeastSmallerString()
Before this change, in StringValue::LeastSmallerString() we did not
check whether '*this' was an empty string and used 'len - 1' as the size
of a new std::string, which could lead to an exception. This change adds
a check for 'len == 0' and we return an empty string in that case.

Also, the function name should actually be LargestSmallerString(), as we
are interested in the largest string that is smaller than '*this', cf.
StringValue::LeastLargerString(). This patch renames the function.

Testing:
 - Expanded the relevant unit test in string-value-test.cc with the test
   of an empty string.

Change-Id: I0888f89070a1efeb9efefd1c3ca96dfa873942fd
Reviewed-on: http://gerrit.cloudera.org:8080/20579
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>