IMPALA-12492: Add metrics for total pending events and lag time of the event-processor
We have last-synced-event-time and latest-event-time on the catalogd
WebUI. It's tedious for users to calculate their difference to get the
lag time. This patch addes the lag in the WebUI directly. The unit is
set to TIME_S so it's shown in a human readable format. Also adds the
total pending events in the same page.

Tests
 - Manually verify the metrics by generating lots of HMS events,
   hard-code the EVENTS_BATCH_SIZE_PER_RPC to 5 and add some sleeps in
   event processing. So it's more easy to show lags.

Change-Id: Id329879bfacba9c8415f920b57939dc571d1aad9
Reviewed-on: http://gerrit.cloudera.org:8080/20560
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>