IMPALA-12483: Add new api in MetastoreServiceHandler for update_transaction_statistics method
In the MetastoreServiceHandler class, the
'update_transaction_statistics' API implementation is missing which is
causing test failures in the external front end integration tests.
This needs to be fixed by adding implementation of the
'update_transaction_statistics' api in MetastoreServiceHandler class.

This API was introduced in HMS long ago but only recently we
encountered a bug with this missing API when testing materialized views
with transactional properties in the external front-end setup.

Testing:
Added a test in the 'test_hms_apis()' to verify that the
'update_transaction_statistics' api in HMS is reachable from impala.

Change-Id: Icd9bc47ff8c6e5e7b31961d7ee6b8c1d4bd74870
Reviewed-on: http://gerrit.cloudera.org:8080/20567
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>