IMPALA-12493: Allow retry after failed cancel_query
Previously if /cancel_query failed with "Query not yet running",
subsequent attempts to cancel the query would always fail and the query
would never be unregistered (so it would stay in "queries in flight" or
"waiting to be closed" until the coordinator was restarted) because the
QueryDriver thought the query was finalized. Check whether the query is
inflight before setting finalized_ to allow retries.

Cancelling a request can no longer fail, so checks for whether the query
is inflight are moved to the callers.

Two ClientRequestState::Finalize calls that previously passed
check_inflight=true no longer check inflight. Retried queries should
always reach inflight before they need to be finalized. They were added
in IMPALA-10413 and IMPALA-10414 with no explanation why check inflight
was needed - presumably it was done because earlier in the function the
retried query was marked inflight - and still pass their tests. Added
DCHECKs since the query should be inflight until closed.

Testing:
- adds a test that canceling a query in CREATED state fails but does not
  prevent canceling the query later
- adds a test that canceling a query that later fails results in that
  query completing rather than going to "waiting to be closed"

Change-Id: I6e2c8872e91a7ff078cedb13e0771bcfaae6ee24
Reviewed-on: http://gerrit.cloudera.org:8080/20584
Tested-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>